refactor: move change stream cursor to separate file
